[PRMT-4566] separate client and repository layer
